Some... modifications to the rules page.
There's a bunch of stuff there that's wrong, outdated or misleading.
It would be nice to update those, or failing that, remove them.
It's just that it needs to be done by someone who actually knows the current state of facts
well enough to confidently document them. That's not me.